last version

+++ NeptuneWebAutheticator/src/main/java/com/NeptuneWebAutheticator/Service/AuthenticationServiceImpl.java
@@ -2,16 +2,21 @@ package com.NeptuneWebAutheticator.Service;
 
 import com.NeptuneWebAutheticator.Entity.Tenant;
 import com.NeptuneWebAutheticator.Entity.Token;
-import com.NeptuneWebAutheticator.Repository.TokenRepository;
 import com.NeptuneWebAutheticator.Repository.TenantRepository;
+import com.NeptuneWebAutheticator.Repository.TokenRepository;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.time.Duration;
 import java.time.Instant;
 import java.time.ZoneId;
@@ -29,10 +34,16 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     @Autowired
     private TokenRepository tokenRepository;
 
+    @Bean
+    public PasswordEncoder passwordEncoderCustom() {
+        return new BCryptPasswordEncoder();
+    }
+
+
     public boolean isTokenValid(Token token) {
         log.info(token.getToken());
-        log.info(""+token.getExpirationDate().toInstant()+ " "+ Instant.now().atZone(ZoneId.of("Europe/Rome")).toInstant());
-        log.info("data token"+token.getExpirationDate().toInstant().isAfter(Instant.now().atZone(ZoneId.of("Europe/Rome")).toInstant()));
+        log.info("" + token.getExpirationDate().toInstant() + " " + Instant.now().atZone(ZoneId.of("Europe/Rome")).toInstant());
+        log.info("data token" + token.getExpirationDate().toInstant().isAfter(Instant.now().atZone(ZoneId.of("Europe/Rome")).toInstant()));
         Instant exp = token.getExpirationDate().toInstant(); // o token.getExpirationDate() se è già Instant
         Instant now = Instant.now();
 
@@ -66,6 +77,15 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     }
 
     @Override
+    public ResponseEntity<String> getUserRole(HttpServletRequest request) {
+        Tenant tenant = getUserAuthenticated(request);
+        if (tenant == null) {
+            return ResponseEntity.status(404).body("Utente non configurato");
+        }
+        return ResponseEntity.ok(tenant.getRole());
+    }
+
+    @Override
     public Tenant getUserAuthenticated(HttpServletRequest request) {
         String session = Arrays.stream(request.getCookies())
                 .filter(c -> "SESSION".equals(c.getName()))
@@ -91,8 +111,14 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         if (isUserAuthenticate(request)) {
             return ResponseEntity.ok().body("login effettuato con successo");
         }
-        Tenant user = userRepository.findByUsernameAndPassword(username, password);
+
+        Tenant user = userRepository.findByUsername(username);
+
         if (user != null) {
+            if (!passwordEncoderCustom().matches(password, user.getPassword())) {
+                return ResponseEntity.status(401).body("Password errata");
+            }
+
             Token token = tokenRepository.findByTenantId(user.getTenantId());
             if (token != null) {
                 tokenRepository.delete(token);
@@ -115,6 +141,25 @@ public class AuthenticationServiceImpl implements AuthenticationService {
                     .header(HttpHeaders.SET_COOKIE, cookie.toString())
                     .body("login effettuato con successo");
         }
-        return ResponseEntity.status(401).body("utente non configurato o credenziali errate");
+        return ResponseEntity.status(401).body("Utente non configurato");
+    }
+
+    @Override
+    @Transactional
+    public ResponseEntity<String> logout(HttpServletRequest request) {
+        if (!isUserAuthenticate(request)) {
+            return ResponseEntity.ok().body("logout effettuato con successo");
+        }
+
+        Tenant tenant = getUserAuthenticated(request);
+
+        if (tenant == null) {
+            return ResponseEntity.status(404).body("Logout per utente non trovato");
+        }
+
+        Token token = tokenRepository.findByTenantId(tenant.getTenantId());
+
+        tokenRepository.delete(token);
+        return ResponseEntity.ok().body("logout effettuato con successo");
     }
 }
